uk-esos-app-web/src/app/payment/shared/pipes/payment-status.pipe.ts
import { Pipe, PipeTransform } from '@angular/core';

import { PaymentCancelledRequestActionPayload, PaymentProcessedRequestActionPayload } from 'esos-api';

@Pipe({ name: 'paymentStatus' })
export class PaymentStatusPipe implements PipeTransform {
  transform(
    value: PaymentProcessedRequestActionPayload['status'] | PaymentCancelledRequestActionPayload['status'],
  ): string {
    switch (value) {
      case 'CANCELLED':
        return 'Cancelled';
      case 'COMPLETED':
        return 'Completed';
      case 'MARK_AS_PAID':
        return 'Marked as paid';
      case 'MARK_AS_RECEIVED':
        return 'Marked as received';
      default:
        return 'Not paid';
    }
  }
}
